[Functional] Migrate backend(Base database) to Firebase Realtime Database

+++ app/src/main/java/com/example/shereats/utils/ImageUtil.kt
@@ -154,7 +154,7 @@ class ImageUtil {
         /**
          * Return the badge color according to its rarity
          */
-        fun getBadgeColor(rarity: Int): Int{
+        fun getBadgeColor(rarity: Long): Int{
             return ConstantUtil.MAP_BADGE_COLOR[rarity] ?: Color.parseColor("#000000")
         }
