+++ app/src/main/java/com/example/shereats/view/adapter/OrderAdapter.kt
@@ -44,7 +44,7 @@ class OrderAdapter(var data: List<Order>): RecyclerView.Adapter<OrderAdapter.Ord
     override fun onBindViewHolder(holder: OrderViewHolder, position: Int) {
         val dataSlice = data[position]
         holder.title.text = dataSlice.id
-        holder.price.text = TextUtil.getItemPrice(dataSlice.price)
+        holder.price.text = TextUtil.getItemPrice(dataSlice.price.toDouble())
         holder.content.text = TextUtil.getOrderContent(dataSlice.items)
         holder.info.text = TextUtil.getOrderInfo(dataSlice)
         holder.contact.setOnClickListener {
